requirements.txt
Appium-Python-Client==0.47
robotframework==3.1.2
selenium==3.141.0
urllib3==1.25.3
